Fix blank facebook response in Public API
Facebook events were showing up in the database but not in the public
API response. This was because the facebook.py connector was setting
is_searchable == is_private within Connector's _convert_event() method.
This was not the intended behavior.

The PublicEventViewSet defined in event_exim/views filters out events
with is_searchable == true and is_private == false. Since facebook.py
always sets both to the same value, this condition could never be true.

This commit updates the facebook.py Connector class to set is_searchable
opposite of is_private. This is the intended behavior, as private events
should be hidden in an API, and public events should be visible in an
API.