Attach event listener to delete buttons and add id to log when created
The delete button has two data attributes:
id: necessary to compose the url and so that the view know what record 
to delete.

click: need to check whether or not an event listener has already been 
attached. The event listener has to be attached on first page load and 
after each new log is created. This is no problem when working with only 
one user record, but when there are multiple user records the event 
listeners end up being attached more than once, firing multiple delete 
requests for a single review log.